Make the DOM NodeId the arena handle itself, dropping the id table
The branch gave every node a monotonically increasing id and mapped it onto
the arena key through a side table, so a freed id could never come back
attached to a stranger. The guarantee held, but it was bought with a
*dependent* load on every tree-walk step: follow a link, index the table,
index the arena. That cost -14.27% across the bench set, and moving the
tree's own links into slot space only took it to -2.24% — the crossing
itself was the price, not where the table sat.

Collapse the two spaces instead of optimising the mapping between them. A
`NodeId` is now the handle: the arena key its state lives at, plus the
generation that key was at when the handle was made. `NodeSlot` becomes an
alias, since a handle and a storage position are the same thing again.

The generation is what keeps the guarantee the counter used to provide.
Freeing bumps a key's generation, so a handle taken before that free no
longer matches and every getter answers `None` rather than whatever moved in
after it. That matters most where script holds ids: a `FinalizationRegistry`
cleanup arriving late carries the handle of the element it was registered
for, and cannot drop the element that took its slot.

Three boundaries can only carry an integer, and they do not all want the same
one. Script's `unique_id` and the image registry take the packed form
(`to_bits`, generation above the key, inside the 2^53 an `f64` represents
exactly). Stylo's `OpaqueNode` takes the bare arena key, which is unique
among simultaneously-live nodes — all it needs — and so still fits a 32-bit
`usize` on wasm32, where the packed form would not.

The deliberate leak goes with the table: nothing is sized by nodes-ever-
created any more. `generations` is indexed by arena key, so it is bounded by
the peak number of live nodes. The document node and document element are
keys 1 and 2 at generation zero and can never be freed, so they still pack to
1 and 2 — an id only becomes a large number once its slot has actually been
recycled.

+++ crates/dom/src/render/images.rs
@@ -19,7 +19,7 @@ use vello::peniko::ImageData;
 #[derive(Debug, Default)]
 pub struct ImageStore {
     by_url: FxHashMap<String, ImageData>,
-    by_node: FxHashMap<usize, ImageData>,
+    by_node: FxHashMap<u64, ImageData>,
 }
 
 impl ImageStore {
@@ -36,7 +36,7 @@ impl ImageStore {
 
     /// Registers the decoded content of a replaced element, replacing any
     /// previous entry.
-    pub fn insert_node(&mut self, node: usize, image: ImageData) {
+    pub fn insert_node(&mut self, node: u64, image: ImageData) {
         self.by_node.insert(node, image);
     }
 
@@ -44,7 +44,7 @@ impl ImageStore {
         self.by_url.remove(url)
     }
 
-    pub fn remove_node(&mut self, node: usize) -> Option<ImageData> {
+    pub fn remove_node(&mut self, node: u64) -> Option<ImageData> {
         self.by_node.remove(&node)
     }
 
@@ -54,7 +54,7 @@ impl ImageStore {
     }
 
     #[must_use]
-    pub fn node(&self, node: usize) -> Option<&ImageData> {
+    pub fn node(&self, node: u64) -> Option<&ImageData> {
         self.by_node.get(&node)
     }
 }
